Avoid locale flash with a client-only editor
